Swap for an external redirect component.
